Added base Collision script
Different types of collision detection can now inherit from the Collision class to already have some base functionality implemented
