+++ Assets/Scripts/Collision/Collision.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c04af13c3898a364ba1ee2945eceb15d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
